python-ecore: Remove unreachable code.
We are not able to reach the retrun with the raise just one line before.
If any python master thinks I'm wrong here please go ahead and revert
and let me know why we would be able to reach it.

This should fix the last warning I'm seeing in the python bindings.

git-svn-id: http://svn.enlightenment.org/svn/e/trunk/BINDINGS/python@79477 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33